refactor: type 패키지 이동 및 회원 등급 enum 변경 (#15)
