+++ src/main/java/com/octagram/pollet/member/domain/model/type/AuthProvider.java
@@ -1,0 +1,5 @@
+package com.octagram.pollet.member.domain.model.type;
+
+public enum AuthProvider {
+	KAKAO, GOOGLE
+}
