+++ src/main/java/com/octagram/pollet/member/domain/model/type/MemberStatus.java
@@ -1,4 +1,4 @@
-package com.octagram.pollet.member.domain.type;
+package com.octagram.pollet.member.domain.model.type;
 
 public enum MemberStatus {
 	ACTIVE, SUSPENDED, BANNED
