+++ src/main/java/com/octagram/pollet/member/domain/model/type/Role.java
@@ -1,0 +1,5 @@
+package com.octagram.pollet.member.domain.model.type;
+
+public enum Role {
+	GUEST, MEMBER, ADMIN
+}
