--- src/main/java/com/octagram/pollet/member/domain/type/AuthProvider.java
@@ -1,5 +1,0 @@
-package com.octagram.pollet.member.domain.type;
-
-public enum AuthProvider {
-	KAKAO, GOOGLE
-}
