--- src/main/java/com/octagram/pollet/member/domain/type/Rank.java
@@ -1,5 +1,0 @@
-package com.octagram.pollet.member.domain.type;
-
-public enum Rank {
-	BRONZE, SILVER, GOLD
-}
